Change: Explicitly type cast inputs to `tau_adaptive`

+++ pyssa/simulation.py
@@ -232,10 +232,10 @@ class Simulation:
                 nc = 10
             for index in range(n_rep):
                 t, X, status = tau_adaptive(
-                    self._react_stoic,
-                    self._prod_stoic,
-                    self._init_state,
-                    self._k_det,
+                    np.int64(self._react_stoic),
+                    np.int64(self._prod_stoic),
+                    np.int64(self._init_state),
+                    np.float64(self._k_det),
                     nc,
                     epsilon,
                     max_t,
